maintainer-scripts: add u-boot install steps for the d1-nezha RiscV board
This includes a new approach to handling opensbi. Don't let the user
manually compile opensbi, but instead require a symlink in the target dir.

That still allows pointing to specific opensbi implementations,
but doesn't require additional manual steps.

Also build opensbi in a per-board build directory, so that we can re-use
the same source repository.

Signed-off-by: Heiko Stuebner <heiko@sntech.de>