maintainer-scripts: allow building new-style u-boot binaries using binman
U-Boot changed the way it builds Rockchip binaries to use binman.
For this the BL31 binary-patch needs to be supplied as BL31 variable
to the build itself, instead of later building a separate uboot.itb

Add support for that while still keeping the old way for now, so that
my old netboots keep compiling until I find time to migrate.

Signed-off-by: Heiko Stuebner <heiko@sntech.de>