maintainer-scripts: more sensible selection of number of compile-threads
Till now the main icecc offload was a Ryzen 3600 (6 cores/12 threads),
so a simple default of 14 compile-threads was not really overloading
the main machine even in the on-the-road case.

With the offloader now being a Ryzen 5900X (12 cores/24 threads) this
changes, as this can easily do 28 compile threads, but would overwhelm
the single machine easily.

So just check the host ip for being in the home network and set compile
threads dynamically if at home or on the road.

Signed-off-by: Heiko Stuebner <heiko@sntech.de>